cluster: rename src/config_manager/ -> src/cluster/
Broader name to reflect the folder's real scope: not just configuration
metadata, but the cluster-level subsystem where topology, sharding
policy, epoch state, node addressing, and their consumer-side cache /
watcher live. Room to grow into membership control and health
monitoring without another rename.

Mechanical rename:
 - git mv src/config_manager -> src/cluster (all 10 files).
 - #include "config_manager/foo.h" -> #include "cluster/foo.h" across
   src/ and tests/ (10 files updated).
 - CMakeLists.txt: DEPTRAN_CORE_SRC glob src/config_manager/*.cc ->
   src/cluster/*.cc; test_config_manager target sources updated;
   stale comments swept.
 - src/deptran/raft/replicated_db.h: stale doc comment referencing
   the old path updated.

Class names (ConfigManager, ClusterConfig, ConfigWatcher, ConfigStore,
ReplicatedKV) stay the same — they describe types, not folder scope,
and renaming them is a separate concern.

test_config_manager still passes 21/21 in <1ms.
