+++ src/cluster/config_manager.cc
