+++ src/cluster/in_memory_replicated_kv.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "config_manager/replicated_kv.h"
+#include "cluster/replicated_kv.h"
 
 #include <map>
 #include <mutex>
